DrawingListFragment 인기순 / 최신순 조회 완료

See merge request s09-webmobile4-sub2/S09P12D101!53

+++ frontend/likloud/app/src/main/java/com/ssafy/likloud/data/model/DrawingListDto.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class DrawingListDto(
 
     @SerializedName("drawingId")
-    val _id: Long,
+    val _id: Int,
 
     @SerializedName("title")
     val title: String,
